requirements.txt
bcrypt==5.0.0
blinker==1.9.0
boto3==1.42.31
botocore==1.42.31
click==8.3.1
colorama==0.4.6
Flask==3.1.2
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-SQLAlchemy==3.1.1
greenlet==3.3.0
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
MarkupSafe==3.0.3
numpy==2.4.1
opencv-python-headless==4.13.0.90
packaging==25.0
pillow==12.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
s3transfer==0.16.0
six==1.17.0
SQLAlchemy==2.0.45
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.5
